Consolitdate and move salt-ssh unit tests to proper location
